refactor: centralize profile default values
